Holy optimizations. 57600 particles with ~1000 FPS

+++ src/state/menu_state.cpp
@@ -58,7 +58,7 @@ MenuState::MenuState(Engine* engine, sf::RenderWindow& window)
     );
 }
 
-void MenuState::update(double dt) {
+void MenuState::update(float dt) {
     position = sf::Mouse::getPosition(window);
 }
