+++ src/state/state.hpp
@@ -9,7 +9,7 @@ public:
      explicit State(Engine* engine) : engine(engine) {}
      virtual ~State() = default;
 
-     virtual void update(double dt) = 0;
+     virtual void update(float dt) = 0;
      virtual void render(sf::RenderWindow& window) = 0;
      virtual void enter() = 0;
      virtual void exit() = 0;
